+++ src/pages/Homepage.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Header from './Header';
-import Body from './Body';
+import Header from '../components/Header';
+import Body from '../components/Body';
 
 function Homepage() {
   return (
